pyproject.toml
[project]
name = "booking-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt==4.3.0",
    "celery==5.3.6",
    "celery[redis]>=5.4.0",
    "fastapi>=0.135.3",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "redis==5.0.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
    "kombu==5.3.5",
]
